fix: initial reset of manual typing date values
When entering the date value to KolInputDate or
number value of KolInputNumber, the first digit
was initial cleared. This happened because the value of the
field was changed from undefined to a blank string. The mapping of the
value from the HTML input tag to the date or number component now
happens only when the value is switched between general blank values
and date or number values.

To prevent stencil warnings, the value props are marked as mutable.

Closes: #2536